Changing lookup table to work
I'm changing how the lookup table works as that gets the output
for the user working. However this is not a good solution.

At issue is the fact that the lookup table to use is happening before
I determine if the identifier is verified or not, and as such the
verification is failing.