Updating exception for reek
As I thought about refactoring the code, I could have added the
explicit args. However that pushed the method length beyond an
acceptable length. The other option was to opt for a dependency
injection container. But that concept neither exists, nor does it
lend itself towards a narrow interface (without creating a custom
parameter object that is used in one case).